\name{cisotope-package}
\alias{cisotope-package}
%%\alias{calc_cisoscape}
\title{
Carbon Isotope and Isoscape Package
}
\description{
This package allows the user to model C3/C4 distribution and create custom carbon isoscapes.
}

\details{
\tabular{ll}{
Package: \tab Cisotope\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-06\cr
License: \tab GPL-2\cr
}
}

\author{
Maintainer: Sydney M. Firmin
}
\references{
Collatz, G.J., Berry, J.A., & Clark, J.S. (1998) Effects of climate and atmospheric CO2 partial pressure on the global distribution of C4 grasses: present, past, and future. Oecologia.

Still, C.J., Berry, J.A., Collatz, G.J. & DeFries, R.S. (2003) Global distribution of C3 and C4 vegetation: carbon cycle implications. Global Biogeochemical Cycles, 17, 1006.

Powell, R.L., Yoo, E-H., & Still, C.J. (2012) Vegetation and soil carbon-13 isoscapes for South America: integrating remote sensing and ecosystem isotope measurements. Ecosphere, 3, 109.
}
\keyword{ carbon-isotope }
