\name{combine_endmembers}
\alias{combine_endmembers}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Combine Endmembers to Calculate Stable Terrestrial Carbon Isoscape
}
\description{
Predicts approximate per-pixel stable carbon isotope ratio for terrestrial landscape based on the percent del13c vegetation life form layers, generated by the del13c function included in the package.
}
\usage{
combine_endmembers(..., landmask, filename = '')
}
\arguments{
  \item{...}{
RasterLayers. del13C layers being combined into final isoscape (output of del13C function). User is able to provide as many layers as they would like. The sum of the original percent vegetation layers the included del13C layers are based on should all sum to 100 percent (representing all components of original landscape)
}
  \item{landmask}{
RasterLayer. Land mask generated from climate data to remove bad data values.
}
  \item{filename}{
Character. Output filename (optional)
}
}
\details{
This function aggregates all of the individual del13C percent plant functional type and vegetation lifeform layers into a final stable terrestrial carbon isoscape. The del13C layers are generated by the PFTdel13C function included in this package, which takes a per-pixel percent plant functional type layer and applies a del13C endmember.
}
\value{
RasterLayer. Final del13C isoscape.
}
\references{
Still, C.J., Berry, J.A., Collatz, G.J. & DeFries, R.S. (2003) Global distribution of C3 and C4 vegetation: carbon cycle implications. Global Biogeochemical Cycles, 17, 1006, doi: 10.1029/2001GB001807.

Powell, R.L., Yoo, E-H., & Still, C.J. (2012) Vegetation and soil carbon-13 isoscapes for South America: integrating remote sensing and ecosystem isotope measurements. Ecosphere, 3, 109, doi: 10.1980/ES12-00162.1.
}
\author{
Sydney M Firmin
}
\examples{
##c4layer <- del13c(total_veg, PFTlayer, endmember, filename)
##c3layer <- del13c(total_veg, PFTlayer, endmember, filename)
##woody <- del13c(total_veg, PFTlayer, endmember, filename)
##finalisoscape <- combine_endmembers(c4layer, c3layer, woody,
##landmask = mask.tif, filename ="isoscape")
}
\keyword{ isoscape }
\keyword{ del13C }
