\name{COMeanppt}
\alias{COMeanppt}
%- Also NEED an '\alias' for EACH other topic documented here.
\docType{data}
\title{
Mean Monthly Precipitation (mm) for Colorado, USA.
}
\description{
RasterBrick. Mean monthly surface precipitation at 10' latitude/longitude spatial resolution cropped to the Spatial extent of Colorado, USA. Interpollated from a data set of station means for the period centered on 1961 to 1990. Precipitation was first defined in terms of the parameters of the Gamma distribution, enabling the calculation of monthly precipitation at any given return period.
}
\usage{data("COMeanppt")}
\format{
Formal class 'RasterBrick' [package "raster"] with 12 slots
}
\source{
http://www.cru.uea.ac.uk/data
}
\references{
New, M., Lister, D., Hulme, M., & Maken, I. (2002) A high-resolution data set of surface climate over global land areas. Climate Research, 21, 1-25.
}
\keyword{datasets}
\keyword{climate}
